requirements.txt
# Python dependencies for Master's AI Assistant

# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
python-multipart==0.0.18

# LangChain & AI
langchain==0.3.14
langchain-groq==0.2.3
langchain-community==0.3.14
langchain-huggingface==0.1.2

# Document Processing
pypdf==5.1.0

# Vector Store & Embeddings
chromadb==0.5.23
sentence-transformers==3.3.1

# Utilities
pydantic==2.10.4
pydantic-settings==2.7.1
